backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
python-docx==1.1.0
pypdf==3.17.1
werkzeug==3.0.1
